+++ notify/schema_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/grafana/alerting/receivers/jira"
 	"github.com/grafana/alerting/receivers/kafka"
 	"github.com/grafana/alerting/receivers/line"
+	"github.com/grafana/alerting/receivers/matrix"
 	"github.com/grafana/alerting/receivers/mqtt"
 	"github.com/grafana/alerting/receivers/oncall"
 	"github.com/grafana/alerting/receivers/opsgenie"
@@ -81,6 +82,7 @@ func TestGetSecretKeysForContactPointType(t *testing.T) {
 		{receiverType: opsgenie.Type, version: schema.V1, expectedSecretFields: []string{"apiKey"}},
 		{receiverType: webex.Type, version: schema.V1, expectedSecretFields: []string{"bot_token"}},
 		{receiverType: sns.Type, version: schema.V1, expectedSecretFields: []string{"sigv4.access_key", "sigv4.secret_key"}},
+		{receiverType: matrix.Type, version: schema.V1, expectedSecretFields: []string{"accessToken"}},
 		{receiverType: mqtt.Type, version: schema.V1, expectedSecretFields: []string{"password", "tlsConfig.caCertificate", "tlsConfig.clientCertificate", "tlsConfig.clientKey"}},
 		{receiverType: jira.Type, version: schema.V1, expectedSecretFields: []string{"user", "password", "api_token"}},
 		{receiverType: victorops.Type, version: schema.V0mimir1, expectedSecretFields: append([]string{"api_key"}, httpConfigSecrets...)},